requirements.txt
Django
django_rest_framework
django-model-utils